services/agents/data_processor/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.5
python-dotenv==1.0.1
boto3==1.35.94
botocore==1.35.94
